pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = ""
authors = ["Matheus Bezerra <fernandesmatheusx12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.40"
asyncpg = "^0.26.0"
python-dotenv = "^0.20.0"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
